pyproject.toml
[tool.poetry]
name = "as_casestudy"
version = "0.1.0"
description = "Post processing for object detection based instance segmentation."
authors = ["Alexander Auer"]

[tool.poetry.dependencies]
python = "3.10.6"
notebook = "^6.4.12"
numpy = "^1.23.2"
opencv-python = "^4.6.0.66"
pandas = "^1.4.4"
Pillow = "^9.2.0"
scikit-image = "^0.19.3"
scipy = "1.9.0"
Shapely = "^1.8.4"

[tool.black]
line-length = 120 # I like to give a bit more space for long lines

[tool.poetry.group.dev.dependencies]
black = "^22.8.0"
flake8 = "^5.0.4"
pytest = "^7.1.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
